Minor Bug fixes
